Added help() and list_exit_codes() functions
For easier customization and code management:
Displaying help and exit codes now invoke functions set up at the
beginning of the script.
Nothing should be change in the CORE section of the script.
The CORE section is only here to run help(), list_exit_codes() or run
main()

The CORE block at the end of the script is no longer to be modified when
adding features. Only the functions at the beginning of the script
should be updated.